backend/src/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==4.0.1
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
openai==1.35.10
pandas==2.2.2
protobuf==4.25.3
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.0
setuptools==70.1.1
tensorflow==2.16.2
urllib3==2.2.2
waitress==3.0.0
Werkzeug==3.0.3